Make more flexible
Any set of codepoints (namelist, pyfontaine, ttf font)
can be used as either a subset (as a namelist usually would
be) or as a full item to check (usually a ttf font).

This refactoring is needed to test what codepoints from a font
(subset) are not supported by a CSS file.